requirements.txt
fastapi
uvicorn
ultralytics
pillow